Exported missing pongoSchema
